scripts/local_test/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
boto3==1.41.5
boto3-stubs==1.41.5
botocore==1.41.5
certifi==2025.11.12
charset-normalizer==3.4.4
duckdb==1.4.2
frozenlist==1.8.0
idna==3.11
jmespath==1.0.1
multidict==6.7.0
polars==1.35.2
polars-runtime-32==1.35.2
propcache==0.4.1
pyarrow==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.15.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
